requirements.txt
flask
requests
flask-cors
gunicorn
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
#cloud-init==21.4
colorama==0.4.3
#command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.11.0
#dbus-python==1.2.16
decorator==4.4.2
distro==1.4.0
#distro-info===0.23ubuntu1
entrypoints==0.3
Flask==1.1.1
fonttools==4.32.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
IPy==1.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.2
#language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
matplotlib==3.5.1
more-itertools==4.2.0
netifaces==0.10.4
networkx==2.4
numpy==1.22.3
oauthlib==3.1.0
packaging==21.3
pandas==1.4.2
pexpect==4.6.0
Pillow==9.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.8
pyrsistent==0.15.5
pyserial==3.4
#python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
#python-debian===0.1.36ubuntu1
pytz==2022.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
scipy==1.8.0
seaborn==0.11.2
SecretStorage==2.3.1
#selinux==3.0
#sepolicy==3.0
service-identity==18.1.0
#setools==4.2.2
simplejson==3.16.0
#six==1.14.0
#sos==4.1
ssh-import-id==5.10
#systemd-python==234
Twisted==18.9.0
#ubuntu-advantage-tools==27.5
#ufw==0.36
#unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
Werkzeug==0.16.1
zipp==1.0.0
#zope.interface==4.7.1
